requirements.txt
fuzzywuzzy==0.18.0
mutagen==1.47.0
pillow==12.2.0
pymediainfo==7.0.1
PyQt6==6.11.0
requests==2.33.1
tinytag==2.2.1